% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Aids2_df}
\alias{Aids2_df}
\title{Australian AIDS Survival Data}
\format{
A data frame with 2843 observations and 7 variables:
\describe{
  \item{state}{A factor indicating the state of residence of the patient (4 levels).}
  \item{sex}{A factor indicating the sex of the patient (2 levels).}
  \item{diag}{An integer indicating the year of diagnosis.}
  \item{death}{An integer indicating the year of death.}
  \item{status}{A factor indicating the status of the patient (2 levels: alive or deceased).}
  \item{T.categ}{A factor indicating the T-cell category of the patient (8 levels).}
  \item{age}{An integer indicating the age of the patient at diagnosis.}
}
}
\source{
Australian Department of Health.
}
\usage{
data(Aids2_df)
}
\description{
The dataset name has been changed to 'Aids2_df' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'HealthDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset provides information on the survival rates and characteristics of AIDS patients in
Australia, including their demographic details and clinical data.
}
