% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{babies_crawl_tbl_df}
\alias{babies_crawl_tbl_df}
\title{Crawling Age}
\format{
A tibble with 12 observations and 5 variables:
\describe{
  \item{birth_month}{A factor indicating the month of birth (January to December).}
  \item{avg_crawling_age}{A numeric value representing the average crawling age (in months).}
  \item{sd}{A numeric value indicating the standard deviation of crawling age.}
  \item{n}{An integer representing the number of infants included in the calculation.}
  \item{temperature}{An integer indicating the average temperature (in degrees Celsius) during the month.}
}
}
\source{
Data collected on the crawling age of infants based on birth month and temperature.
}
\usage{
data(babies_crawl_tbl_df)
}
\description{
The dataset name has been changed to 'babies_crawl_tbl_df' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_tbl_df' indicates that this dataset is a tibble, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains information on the average crawling age of infants based on the month of birth,
as well as associated factors such as temperature.
}
