% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{bdims_tbl_df}
\alias{bdims_tbl_df}
\title{Body measurements of 507 physically active individuals.}
\format{
A tibble with 507 observations and 25 variables:
\describe{
  \item{bia_di}{Numerical value of body measurement (in cm).}
  \item{bii_di}{Numerical value of body measurement (in cm).}
  \item{bit_di}{Numerical value of body measurement (in cm).}
  \item{che_de}{Numerical value of body measurement (in cm).}
  \item{che_di}{Numerical value of body measurement (in cm).}
  \item{elb_di}{Numerical value of body measurement (in cm).}
  \item{wri_di}{Numerical value of body measurement (in cm).}
  \item{kne_di}{Numerical value of body measurement (in cm).}
  \item{ank_di}{Numerical value of body measurement (in cm).}
  \item{sho_gi}{Numerical value of body measurement (in cm).}
  \item{che_gi}{Numerical value of body measurement (in cm).}
  \item{wai_gi}{Numerical value of body measurement (in cm).}
  \item{nav_gi}{Numerical value of body measurement (in cm).}
  \item{hip_gi}{Numerical value of body measurement (in cm).}
  \item{thi_gi}{Numerical value of body measurement (in cm).}
  \item{bic_gi}{Numerical value of body measurement (in cm).}
  \item{for_gi}{Numerical value of body measurement (in cm).}
  \item{kne_gi}{Numerical value of body measurement (in cm).}
  \item{cal_gi}{Numerical value of body measurement (in cm).}
  \item{ank_gi}{Numerical value of body measurement (in cm).}
  \item{wri_gi}{Numerical value of body measurement (in cm).}
  \item{age}{An integer indicating the age of the individual (in years).}
  \item{wgt}{A numeric value indicating the weight of the individual (in kg).}
  \item{hgt}{A numeric value indicating the height of the individual (in cm).}
  \item{sex}{An integer indicating the sex of the individual (1 = male, 2 = female).}
}
}
\source{
Data collected from physically active individuals to analyze body measurements.
}
\usage{
data(bdims_tbl_df)
}
\description{
The dataset name has been changed to 'bdims_tbl_df' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_tbl_df' indicates that this dataset is a tibble, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains body measurements of 507 physically active individuals, including various
dimensions and physical attributes.
}
