% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{colmozzie_tbl_df}
\alias{colmozzie_tbl_df}
\title{Weekly Notified Dengue Cases in Sri Lanka}
\format{
A tibble with 279 observations and 12 variables:
\describe{
  \item{Cases}{Number of dengue cases reported during the week (integer).}
  \item{Year}{Year of the reported cases (integer).}
  \item{Week}{Week of the year (integer).}
  \item{TEM}{Average temperature (numeric).}
  \item{TMAX}{Maximum temperature recorded (numeric).}
  \item{Tm}{Minimum temperature recorded (numeric).}
  \item{SLP}{Sea level pressure (character).}
  \item{H}{Humidity levels (numeric).}
  \item{PP}{Precipitation levels (numeric).}
  \item{VV}{Wind velocity (numeric).}
  \item{V}{Another wind variable (numeric).}
  \item{VM}{Yet another wind variable (numeric).}
}
}
\source{
The dataset is based on public health records and meteorological data from Sri Lanka.
}
\usage{
data(colmozzie_tbl_df)
}
\description{
The dataset name has been changed to 'colmozzie_tbl_df' to avoid confusion with other datasets
from packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_tbl_df' indicates that this dataset is a tibble, helping to distinguish it from other
datasets within the package and from those in the broader R ecosystem. The original content of the
dataset has not been modified in any way.
}
\details{
This dataset contains weekly reported cases of dengue fever in Sri Lanka, along with various
meteorological variables that may be associated with the incidence of the disease.
}
