% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Cushings_df}
\alias{Cushings_df}
\title{Diagnostic Tests on Patients with Cushing's Syndrome}
\format{
A data frame with 27 observations and 3 variables:
\describe{
  \item{Tetrahydrocortisone}{A numeric vector representing the levels of Tetrahydrocortisone.}
  \item{Pregnanetriol}{A numeric vector representing the levels of Pregnanetriol.}
  \item{Type}{A factor indicating the type of test conducted (4 levels).}
}
}
\source{
Data collected from clinical trials and patient records.
}
\usage{
data(Cushings_df)
}
\description{
The dataset name has been changed to 'Cushings_df' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains results from diagnostic tests conducted on patients suspected of having
Cushing's Syndrome, focusing on the measurement of specific hormonal metabolites.
}
