% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{esoph_df}
\alias{esoph_df}
\title{Smoking, Alcohol and (O)esophageal Cancer}
\format{
A data frame with 88 observations and 5 variables:
\describe{
  \item{agegp}{Age group of the individuals (ordered factor).}
  \item{alcgp}{Alcohol consumption group (ordered factor).}
  \item{tobgp}{Tobacco consumption group (ordered factor).}
  \item{ncases}{Number of cases (numeric).}
  \item{ncontrols}{Number of controls (numeric).}
}
}
\source{
Data from a case-control study on esophageal cancer.
}
\usage{
data(esoph_df)
}
\description{
The dataset name has been changed to 'esoph_df' to avoid confusion with datasets from other packages
in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package. The suffix '_df'
indicates that this dataset is a data frame, helping to distinguish it from other datasets within the package
and from those in the broader R ecosystem. The original content of the dataset has not been modified in any way.
}
\details{
This dataset contains data from a case-control study investigating the association between smoking, alcohol consumption,
and esophageal cancer. It includes the number of cancer cases and controls for various age, alcohol consumption, and smoking groups.
}
