% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{fdeaths_ts}
\alias{fdeaths_ts}
\title{Monthly Deaths from Lung Diseases in the UK (Females)}
\format{
A time series object with 72 observations, from 1974 to 1980:
\describe{
  \item{fdeaths}{A numeric vector representing the number of monthly deaths due to lung diseases in females.}
}
}
\source{
UK Health Authority data on lung disease deaths.
}
\usage{
data(fdeaths_ts)
}
\description{
The dataset name has been changed to 'fdeaths_ts' to avoid confusion with datasets from other packages
in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package. The suffix '_ts'
indicates that this dataset is a time series object, helping to distinguish it from other datasets within the package
and from those in the broader R ecosystem. The original content of the dataset has not been modified in any way.
}
\details{
This dataset contains the number of monthly deaths from lung diseases in the UK, specifically for females,
from 1974 to 1980.
}
