% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{GAGurine_df}
\alias{GAGurine_df}
\title{Level of GAG in Urine of Children}
\format{
A data frame with 314 observations and 2 variables:
\describe{
  \item{Age}{A numeric vector representing the age of the children in years.}
  \item{GAG}{A numeric vector representing the levels of glycosaminoglycans (GAG) in urine (in some
  appropriate unit).}
}
}
\source{
Data collected from pediatric health studies.
}
\usage{
data(GAGurine_df)
}
\description{
The dataset name has been changed to 'GAGurine_df' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'ChildHealthDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains measurements of glycosaminoglycan (GAG) levels in urine samples collected
from children, aimed at studying potential health implications associated with abnormal GAG levels.
}
