% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{genotype_df}
\alias{genotype_df}
\title{Rat Genotype Data}
\format{
A data frame with 61 observations and 3 variables:
\describe{
  \item{Litter}{A factor indicating the litter group (with 4 possible levels).}
  \item{Mother}{A factor indicating the mother of the rats (with 4 possible levels).}
  \item{Wt}{A numeric value representing the weight of the rats (in grams).}
}
}
\source{
Data collected from genetic studies involving rat populations.
}
\usage{
data(genotype_df)
}
\description{
The dataset name has been changed to 'genotype_df' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains genotype data from rats, including information on litter, maternal lineage,
and weight measurements.
}
