% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Indometh_df}
\alias{Indometh_df}
\title{Pharmacokinetics of Indomethacin}
\format{
A data frame with 66 observations and 3 variables:
\describe{
  \item{Subject}{Identifier for the subject (ordinal factor with 6 levels).}
  \item{time}{Time since drug administration (numeric, in hours).}
  \item{conc}{Indomethacin concentration (numeric, in mcg/ml).}
}
The dataset contains additional metadata:
\describe{
  \item{formula}{The formula representing concentration over time, grouped by subject: \code{conc ~ time | Subject}.}
  \item{labels}{List containing labels for time and concentration:
  \itemize{
    \item{\code{x}: "Time since drug administration" (hr)}
    \item{\code{y}: "Indomethacin concentration" (mcg/ml)}
  }}
}
}
\source{
Pharmacokinetic study data for Indomethacin.
}
\usage{
data(Indometh_df)
}
\description{
The dataset name has been changed to 'Indometh_df' to avoid confusion with datasets from other
packages in the R ecosystem and to follow the naming convention in the 'MedDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, distinguishing it both within the
package and from similar datasets in other R packages. The original content of the dataset has not been
modified in any way.
}
\details{
This dataset contains pharmacokinetic data for Indomethacin, a nonsteroidal anti-inflammatory drug (NSAID).
It includes concentration measurements over time for multiple subjects, allowing for the analysis of the drug's
pharmacokinetics across individuals.
}
