% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ldeaths_ts}
\alias{ldeaths_ts}
\title{Monthly Deaths from Lung Diseases in the UK}
\format{
A time series object with 72 observations:
\describe{
  \item{ldeaths}{A numeric vector containing the number of monthly deaths from lung diseases in the UK.}
}
}
\source{
Office for National Statistics, UK.
}
\usage{
data(ldeaths_ts)
}
\description{
The dataset name has been changed to 'ldeaths_ts' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_ts' indicates that this dataset is a time series, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset provides information on the monthly deaths from lung diseases in the UK, recorded
from 1974 to 1980.
}
