% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{mala_df}
\alias{mala_df}
\title{The Relationship Between Gene and Disease}
\format{
A data frame with 241,306 observations and 2 variables:
\describe{
  \item{disease}{Disease associated with the gene (factor).}
  \item{gene}{Gene associated with the disease (factor).}
}
}
\source{
The dataset contains gene-disease relationship data from various scientific studies and databases.
}
\usage{
data(mala_df)
}
\description{
The dataset name has been changed to 'mala_df' to avoid confusion with other datasets
from packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, helping to distinguish it from other
datasets within the package and from those in the broader R ecosystem. The original content of the
dataset has not been modified in any way.
}
\details{
This dataset contains information about the relationships between genes and diseases, providing
insights into how specific genes are associated with various diseases. It offers a comprehensive
view of gene-disease associations.
}
