% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meddatasets-package.R
\docType{package}
\name{MedDataSets}
\alias{MedDataSets}
\alias{MedDataSets-package}
\title{MedDataSets: Comprehensive Medical, Disease, Treatment, and Drug Datasets}
\description{
{
The MedDataSets package provides an extensive collection of datasets related to medicine, diseases, treatments, drugs, and public health It covers topics such as pharmacokinetics, drug effectiveness, vaccine trials, survival rates, infectious disease outbreaks, and medical treatments.
}
}
\details{
MedDataSets: Comprehensive Medical, Disease, Treatment, and Drug Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Comprehensive Medical, Disease, Treatment, and Drug Datasets
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/meddatasets}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
