% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{muscle_df}
\alias{muscle_df}
\title{Effect of Calcium Chloride on Muscle Contraction in Rat Hearts}
\format{
A data frame with 60 observations and 3 variables:
\describe{
  \item{Strip}{A factor indicating the specific muscle strip (with 21 possible levels).}
  \item{Conc}{A numeric value representing the concentration of calcium chloride (in mM).}
  \item{Length}{A numeric value indicating the length of the muscle strip (in millimeters).}
}
}
\source{
Data collected from experiments assessing the impact of calcium chloride on muscle contraction.
}
\usage{
data(muscle_df)
}
\description{
The dataset name has been changed to 'muscle_df' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains experimental data on the effect of calcium chloride on muscle contraction
in rat hearts, including measurements of muscle strip length and calcium concentration.
}
