% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Puromycin_df}
\alias{Puromycin_df}
\title{Reaction Velocity of an Enzymatic Reaction}
\format{
A data frame with 23 observations and 3 variables:
\describe{
  \item{conc}{Substrate concentration (numeric).}
  \item{rate}{Reaction velocity (numeric).}
  \item{state}{A factor with two levels: \code{treated} and \code{untreated}, indicating whether Puromycin was present.}
}
The dataset contains additional metadata:
\describe{
  \item{reference}{The reference for this dataset: \code{"A1.3, p. 269"}.}
}
}
\source{
Experimental data on the effects of Puromycin on enzyme reaction rates.
}
\usage{
data(Puromycin_df)
}
\description{
The dataset name has been changed to 'Puromycin_df' to avoid confusion with datasets from other
packages in the R ecosystem and to follow the naming convention in the 'MedDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, distinguishing it both within the
package and from similar datasets in other R packages. The original content of the dataset has not been
modified in any way.
}
\details{
This dataset contains data from an experiment on the reaction velocity of an enzymatic reaction
with and without the use of Puromycin, an antibiotic that inhibits protein synthesis.
}
