% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{sleep_deprivation_tbl_df}
\alias{sleep_deprivation_tbl_df}
\title{Survey on Sleep Deprivation and Transportation Workers}
\format{
A tibble with 1,087 observations and 2 variables:
\describe{
  \item{sleep}{A factor indicating the reported sleep deprivation level (e.g., low, moderate, high).}
  \item{profession}{A factor indicating the profession of the participants (e.g., truck driver, pilot, etc.).}
}
}
\source{
Data collected from a survey targeting transportation workers to assess the impact of sleep
deprivation on their performance and well-being.
}
\usage{
data(sleep_deprivation_tbl_df)
}
\description{
The dataset name has been changed to 'sleep_deprivation_tbl_df' to avoid confusion with other datasets from
packages in the R ecosystem and to follow the naming conventions of the 'MedDataSets' package.
The suffix '_tbl_df' indicates that this dataset is a tibble, helping to distinguish it from
other datasets within the package and from those in the broader R ecosystem. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains information from a survey conducted on transportation workers, focusing on
the relationship between sleep deprivation and their professional roles. It includes variables
on the amount of sleep reported and the professions of the respondents.
}
