% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Theoph_df}
\alias{Theoph_df}
\title{Pharmacokinetics of Theophylline}
\format{
A data frame with 132 observations and 5 variables:
\describe{
  \item{Subject}{An ordered factor indicating the subject ID (12 levels).}
  \item{Wt}{Weight of the subject in kilograms (numeric).}
  \item{Dose}{Theophylline dose administered in milligrams (numeric).}
  \item{Time}{Time since drug administration in hours (numeric).}
  \item{conc}{Theophylline concentration in the blood in mg/L (numeric).}
}
The dataset includes additional metadata:
\describe{
  \item{x}{Time since drug administration, in hours (\code{hr}).}
  \item{y}{Theophylline concentration in mg/L (\code{mg/L}).}
}
}
\source{
Experimental data on the pharmacokinetics of Theophylline.
}
\usage{
data(Theoph_df)
}
\description{
The dataset name has been changed to 'Theoph_df' to avoid confusion with datasets from other
packages in the R ecosystem and to adhere to the naming conventions within the 'MedDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, distinguishing it from other datasets
within the package and from datasets in the broader R ecosystem. The original content of the dataset
has not been modified in any way.
}
\details{
This dataset contains pharmacokinetic data on the drug Theophylline, which is used to treat respiratory diseases.
The dataset includes measurements of drug concentration over time after an oral dose of the drug.
}
