% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ToothGrowth_df}
\alias{ToothGrowth_df}
\title{The Effect of Vitamin C on Tooth Growth in Guinea Pigs}
\format{
A data frame with 60 observations and 3 variables:
\describe{
  \item{len}{Tooth length (numeric).}
  \item{supp}{Type of supplement: either "VC" (Vitamin C) or "OJ" (Orange Juice) (factor).}
  \item{dose}{Dose of Vitamin C administered in milligrams per day (numeric).}
}
}
\source{
Experimental data on the effect of Vitamin C on tooth growth in guinea pigs.
}
\usage{
data(ToothGrowth_df)
}
\description{
The dataset name has been changed to 'ToothGrowth_df' to avoid confusion with datasets from other
packages in the R ecosystem and to align with the naming conventions of the 'MedDataSets' package.
The suffix '_df' indicates that this dataset is a data frame, helping to distinguish it from other
datasets within the package and from datasets in the broader R ecosystem. The original content of the dataset
has not been modified in any way.
}
\details{
This dataset explores the effect of Vitamin C on tooth growth in guinea pigs. It includes data on
tooth length as a response to different doses of Vitamin C, administered through two delivery methods.
}
